requirements.txt
streamlit
python-dotenv
langchain
langchain-groq
fpdf
